refactor(protocol): move the destination peer id parser to address-filter
